+++ packages/search/src/index.d.ts
@@ -40,6 +40,8 @@ export interface SearchPlugin extends Plugin {
     Search: (props: SearchProps) => React.ReactElement;
     ShowSearchPopover: (props: ShowSearchPopoverProps) => React.ReactElement;
     ShowSearchPopoverButton(): React.ReactElement;
+    clearHighlights(): void;
+    highlight(keyword: SingleKeyword | SingleKeyword[]): void;
 }
 
 export type SingleKeyword = string | RegExp;